skip Docker build and comments

+++ camera-app/build.sh
@@ -7,6 +7,9 @@
 #
 # User must have access to Docker.
 #
+# Usage: ./build.sh [-skipDocker]
+#  - with -skipDocker you can skip the Docker build
+# 
 
 NAME=camera-app
 TAG=LATEST
@@ -20,14 +23,19 @@ echo Building the app
 npm install
 ng build --prod
 
-# Remove unused images
-echo Remove unused images
-docker image prune -a -f
+# Building Docker
+if [ "$1" != "-skipDocker" ]
+    then
+        # Remove unused images
+        echo Remove unused images
+        docker image prune -a -f
 
-# Building the image
-echo Building the image
-docker build -t $NAME:$TAG .
+        # Building the image
+        echo Building the image
+        docker build -t $NAME:$TAG .
 
-# Running the container
-#echo Running the container on port 8888
-#docker run --rm -it -p 8888:80 $NAME:$TAG 
+        # Running the container
+        # Container needs the camera-service as linked container with name apiserver
+        #echo Running the container on port 8888
+        #docker run --rm -it --name camera-app --link camera-service:apiserver -p 8888:80 $NAME:$TAG 
+fi
